renamed configureScript to specify yaml because im about to add another configure but for xml
